refactor: input 배열의 유효성 체크하는 util 함수 분리
